fix(tombtweaks) 1.5.1: read enchantment ids as int, not short
The Curse of Possession scan resolved enchantment ids out of `ench` /
`StoredEnchantments` with getShort("id"). JustEnoughIDs widens enchantment ids
from short to int and rewrites vanilla's own reads to match, but it does not
touch ours - so any id above 32767 came back truncated and the scan would either
miss the curse or match the wrong enchantment.

Latent today - the pack is a few hundred enchantments deep, well under the
limit - so this is hardening, not a live fix.

NBTTagCompound.getInteger accepts any numeric tag, so a short-typed "id" still
reads correctly and the change is safe without JEID installed. "lvl" stays a
short deliberately: JEID widens only the id.

+++ tombtweaks/src/main/java/com/spege/tombtweaks/events/CurseOfPossessionHandler.java
@@ -58,7 +58,10 @@ public class CurseOfPossessionHandler {
         if (enchantTag != null) {
             for (int i = 0; i < enchantTag.tagCount(); i++) {
                 NBTTagCompound enchant = enchantTag.getCompoundTagAt(i);
-                Enchantment enchantObj = Enchantment.getEnchantmentByID(enchant.getShort("id"));
+                // getInteger, not getShort: JustEnoughIDs widens enchantment ids past 32767 and
+                // rewrites vanilla's own reads, but not ours. getInteger reads a short tag fine
+                // (NBTTagCompound.getInteger accepts any numeric tag), so this is safe without JEID.
+                Enchantment enchantObj = Enchantment.getEnchantmentByID(enchant.getInteger("id"));
                 if (enchantObj != null) {
                     net.minecraft.util.ResourceLocation regName = enchantObj.getRegistryName();
                     String registryName = regName != null
